Move TodoItem directory and files to client/containers folder
Also added an Edit Form button into PriorityDisplay
